requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via
    #   pyogrio
    #   pyproj
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via flask
colorama==0.4.6
    # via click
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.5
    # via google-auth
cycler==0.12.1
    # via matplotlib
flask==3.1.3
    # via velomunicipal (pyproject.toml)
fonttools==4.62.1
    # via matplotlib
geopandas==1.1.3
    # via velomunicipal (pyproject.toml)
google-api-core==2.30.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.49.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-storage==3.10.0
    # via velomunicipal (pyproject.toml)
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.73.0
    # via google-api-core
idna==3.11
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
kiwisolver==1.5.0
    # via matplotlib
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.10.8
    # via velomunicipal (pyproject.toml)
numpy==2.4.3
    # via
    #   contourpy
    #   geopandas
    #   matplotlib
    #   numpy-typing-compat
    #   optype
    #   pandas
    #   pyogrio
    #   scipy
    #   shapely
numpy-typing-compat==20251206.2.4
    # via optype
optype==0.17.0
    # via scipy-stubs
packaging==26.0
    # via
    #   geopandas
    #   matplotlib
    #   pyogrio
pandas==3.0.1
    # via geopandas
pillow==12.1.1
    # via matplotlib
proto-plus==1.27.1
    # via google-api-core
protobuf==6.33.6
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pyogrio==0.12.1
    # via geopandas
pyparsing==3.3.2
    # via matplotlib
pyproj==3.7.2
    # via geopandas
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
requests==2.32.5
    # via
    #   velomunicipal (pyproject.toml)
    #   google-api-core
    #   google-cloud-storage
scipy==1.17.1
    # via velomunicipal (pyproject.toml)
scipy-stubs==1.17.1.2
    # via velomunicipal (pyproject.toml)
shapely==2.1.2
    # via geopandas
six==1.17.0
    # via python-dateutil
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
werkzeug==3.1.6
    # via flask
